2021.09.15 - Homework

+++ module-1/03_Expressions/lecture-student/dotnet/Lecture/04_ReturnDoubleOfTwo.cs
@@ -6,9 +6,9 @@
          4. This method currently returns an int. Change it so that it returns a double.
             TOPIC: Return Types
         */
-        public int ReturnDoubleOfTwo()
+        public double ReturnDoubleOfTwo()
         {
-            return 2;
+            return 2.0;
         }
     }
 }
